※ 포트폴리오 활용도 있음.  싱글 콘텐츠 쓰레드 모든 lco

+++ _lib/AcceptEx_IOCP_NetworkLib/Session.h
@@ -10,7 +10,7 @@
 namespace network
 {
 using ull = unsigned long long;
-using seqAddrType = long long;
+using seqAddrType = __int64;
 
 struct SeqAndIdx
 {
